Add files via upload
Updates of the calculation of solar and wind capacity:

(1) For both wind and solar calculations, added functions "check_leap_year" to determine if the input year is a leap year or not (used to determine the );
(2) For solar, added the variable "total_sunlight_tolerance". This is used to scale the maximum allowable surface radiation; 
(3) For solar, modified the function "cal_incidence_angles". Panel tilt is now returned to calculate the adjusted diffuse sunlight;